HCF-348 Skip license verification for releases with missing license archives
Bosh's ReleaseArchiver explicitly extracts license.tgz when building
release archives.  Don't error out if license.tgz is missing, even if there
is a SHA1 checksum in the release manifest (because we can't reproducibly
reconstruct the tar even if we wanted to).  It is unclear why the release
archives from bosh.io _do_ (sometimes) include a license.tgz instead of its
component files.