Do not fail if taking ownership fails
Taking ownership of the compiled artifacts can fail depending on the underlying filesystem. An example where this is the case is vmhgfs, where this command fails because the host of the VM may not have a user with the UID and GID that match the user that ran the commands in the guest.
In this case, vmhgfs actually does the UID and GID translation for us, so the chown would be superflous.