HCF-313 - With debug mode, retain the contain only if the command failed with -d given.
Note that we don't do this for build-base, as there's not much to find in
the container if apt-get fails.  Container retention is done by making
the primary command 'sleep 12h' and then running the actual command with
'docker exec'.  This lets us remove the container later if there's no
need to keep it.