HCF-424 - Prevent double Printf-like handling.
%s was interpreted by color.GreenString, inserting the color.RedString(...).
The result going through another Printf-like function (c.ui.Println), as its format-argument then causes mis-interpretation of any %-character in the scanner.Text string.

With the change the color.GreenString is the format-argument of Println, and color.RedString becomes the string argument to insert.
The %s becomes a %%s to prevent GreenString from interpretation. Println then sees the proper %s in its format.

compilator tests pass. Checking ... Yes, output is discarded by the tests, so the changes in them are not visible to the tests.