HCF-344 - First step towards DRY for the fissile cli definitions.
Notes: Many repeated options still come in multiple variants.
List:
  base-image
  release
  repository
  roles-manifest (**)
  target
  prefix (**)
  compiled-packages (**)
  default-consul-address (**)
  default-config-store-prefix (**)
  light-opinions (**)
  dark-opinions (**)
  workers (**)

Some only differ in the (non)use of a fallback environment variable (**),
or of a default value (vs not).

For (**) I would like to know from somebody knowing the fissile commands much better if we can simply switch to the variant with fallback to environment variable in all places, or at least in some.

One possible bug was seen, a possible copy/paste error.
"--target"
  targetConfigEnvFlag
  Described as output path for config files.
  Default value is path to "dockerfiles".
Likely copied from the definition of "targetDocker2EnvFlag".