HCF-583 Reduce copying when building packages layer image
Instead of copying all the compiled files to a scratch workspace and
telling docker to build from there, just build the tar file ourself
(which is what would have happened before too).  This reduces a set of
read/write that didn't need to happen.

In tests, it reduced rebuilding the packages layer image from ~4:30
to ~3:40.