HCF-1057 - Fix leaked config value by copying hashes during flattening
In the test suite, there are two releases, tor and ntpd, that both declare
the tor.private_key field.  This field is non-empty in the ntpd release
but empty in the tor release (it took me a while to memorize this, as you'd
expect the opposite, but no bug happens that way). The reason for that is when
a subtree is copied from the source to the dest, if a later recursive call
modifies that subtree, it modifies the original tree of properties.

The solution is to copy over each subtree on assignment.  No data is ever
copied over more than once -- we either ignore a comparison, assign
the source to the dest, or recurse into both (when they're hashes),
and continue the merge on the children.