[HCF-1155][HCF-776] Implement support for post-start scripts.
At the core it uses a dynamically created monit job which checks that
all the original jobs are up before it then invokes the post-start
scripts of the role to complete itself.

Note, this manual dependency checking is __only__ done because the
available monit (v5.6) has bugs in its support for dependencies. When
a newer monit becomes available the dependency checking can be moved
out of the job start command/script into the job definition, making
monit do that part of the work.