Make compile-without-docker detection stricter
We need to use the fissile-built compilation image to actually build
anything; however, that image already has a `/fissile-in` directory
(because that's how the script to build the image was dropped in there).
So we need to actually check for `/fissile-in/var/vcap` to know if we're
building in docker.