Compute an aggregated signature (spec and opinion values, role version), making the image version dependent on all that input. This is similar to the code in NewDockerPopulator() and WriteConfigs() to write the information into the tar stream. A relevant job method had to be made public.
Notable differences: We flatten the properties from nested map to
simple k/v map. to sort the keys, making the hashing stable. Related
to that is our use of the extended total flattening to see the maps
hidden in array values and allow for them to be ordered as well.

This way the detector reacts to actual content changes, without
triggering due to go's map randomization or simple things like
reordering input without actual changed values.