GG2BiKoZ Support role affinity
Affinity rules can be very complex, so we just copy them verbatim from
the role manifest, without defining any schema for them.

In kube 1.5 they are stored in JSON format as Pod annotations, and in
1.6+ they have their own "spec/affinity" section.

For non-templatized kube configs we use the 1.5 style annotation because
we already use the 1.5 apiVersion for Job objects elsewhere.