S4mXqwBB Append Release.Revision to Job names
We want jobs to run again during `helm upgrade`. This means we have
to delete the job before creating it again, which helm doesn't support,
except for helm "hooks".  Using a random suffix creates a new Job
that will still run while the old one can be removed because it is no
longer included in the chart.

Also create a RELEASE_REVISION environment variable and append the
revision to the secret-update name so the secrets generator can
access the correct generation of this data.