Add a --output-graph option to dump graphviz file for hashes involved
This is to help debugging cases where things were being unexpectedly
rebuilt (or not being rebuilt when it should be).

Also adds a bin/graph-diff to examine the differences between two such
graphs, colour-coding the differences.

The best way to look at the graphs are probably to generate them via
`fdp` and view them with `ZGRViewer`.