Move MakeBasicValues from helm to kube package
The helm package is just a specialized templating library; default values
for kube specs do not belong there but should be with the related functions
in the kube package.

The helm_helper functions are calling MakeBasicValues(), so move that file
into the kube package as well to avoid circular dependencies (it is also
not being used outside the kube package).