[#PMrNrMOh] Move writing of service accounts and roles bound to a single instance group from separate files to the file for the instance group.
This change results in fissile generating a less cluttered helm
chart. Accounts and roles bound just to things like a test task
(irrelevant to helm), for example, will not appear in the helm chart
any longer.

The effect is implemented by fissile maintaining usage/reference
counters for accounts and roles when iterating over instance groups
during load.  When writing files the accounts and roles with a single
user are ignored by the code creating separate files, and the code
writing instance groups is extended to take over that task instead.