Call make/bindata whenever required
This makes it possible to run all the build steps of `make all` in
arbitrary order, e.g. `make clean vet`, `make clean test` etc all
work now.

Running `make all` will now result in running make/bindata multiple
times. But given that it takes just a couple of milliseconds this
doesn't really matter.

The docker images are not required for any of the other build steps,
so they are no longer pulled as part of the `make all` command.