Use legacy rolename label skiff-role-name for selectors
Older releases _only_ used the skiff-role-name label and had no selectors
for stateful sets. Kube would then autogenerate a selector that includes
all the existing labels on the set (which was just skiff-role-name).

By using an identical selector we will still be able to upgrade the
stateful sets (because changes to the spec are not allowed).