diff: dig into underlying types when stringifying
When doing a `fissile diff`, we rely on the stringification of the specs
to be consistent.  However, when we have nested things (hashes-in-arrays
and that kind of things), we end up getting an object of type Interface
instead of the underlying type.  Explicitly unwrap it to ensure we get
the correct results.

The test cases have been updated to attempt to catch this case; however
it is inherently random, so this may occasionally pass even when wrong.