Add support for reverse consumer information
In /var/vcap/job-src/*/config_spec.json, we will now include a list of
consumers of this job's links (via the `consumed_by` key, similar to the
existing `consumes` key).  This is useful so we can attempt to restart
pods which contain outdated information (in case they linked against a
previous iteration of this pod and now contains outdated information).