Fix colocated container volume count validation
Fix the colocated container volume validation that checks for unsed volumes
defined in the main instance group, which are not used by any colocated
container instance group. Do not compare the number of volumes, but check
that each volume definition of the main instance group is at least used
once by one of the colocated containers. The previous logic created false
validation errors in case multiple colocated containers were defined, but
not all of them used a shared volume.