Add feature flag details to values.yaml comments
There are *no* functional changes; all changes only affect the comments
inside values.yaml.

For each instance group that has a feature flag condition this adds a
comment to the sizing.INSTANCEGROUP.count field:, e.g.

sizing:
  diego_api:
    ...
    # The diego_api instance group can be disabled by the eirini feature.
    # It can scale between 1 and 3 instances.
    # For high availability it needs at least 2 instances.
    count: 1

Each feature flag is also annotated with the list of instance groups it
controls, e.g.

enable:
  # The eirini feature enables these instance groups: eirini
  # It disables these instance groups: diego_api, diego_brain, diego_ssh, and
  # diego_cell
  eirini: false

Additional minor changes are adding the conjunctive "and" in the list of 3
or more jobs, and using the proper format for the instance group names
(underscores instead of dashes, 'credhub_user' instead of 'credhub-user').