Remove the helm.Match() methods
They were added in 678b0ba, but are overly complex, given that the
only usage was locating specific resource kinds in a list of nodes.
The list version of Match() also did not actually work (which didn't
matter, as it wasn't used anywhere).

This commit provides a much simplified version of matchNodeInlist()
and gets rid of all the Match() methods. This restores helm/config.go
to 100% test coverage, which was the original motivation for this
change. :)