Allow sizing.XXX.count under config.HA=true to be set to the default SA value
This changes all counts in values.yaml to nil instead of the default SA
(min) value. That way the template can differentiate between the count
being set by the user vs. still having the default value from values.yaml.

The templates will still throw an error when the config.HA is true and
the count for a role is set to less than the ha minimum. If that is
intended, then the new config.HA_strict variable must be set to false.

--set config.HA=true --set congif.HA_strict=false --set sizing.FOO.count=1