Implement "ignore" option for link consumers
In the rolemanifest a job can now explicitly state that a specified bosh
link will *not* be consumed:

  jobs:
  - name: sample-job
    release: sample
    consumes:
      not-used: {ignore: true}

The link definition will not be included in the job information passed on
to Configgin, so the link will not resolve at runtime. It is the responsibility
of the templates to not reference it (by setting properties, or by patching
the templates).

Note that this feature allows ignoring of non-optional links as well!

[jsc#CAP-150]