Implement configgin startup synchronization
The next version of configgin will export bosh properties to a
secret and tag it with the value of CONFIGGIN_VERSION_TAG (the
helm chart version plus the secrets rotation count).

Fissile now defines an invironment variable that references that
role's secret tag in each role that needs to import links from it.
This has the effect of delaying startup of the pod until all links
that have to be imported have been written to their corresponding
secrets.

[jsc#CAP-150]