Define CONFIGGIN_SA_TOKEN for all pods using configgin
It points to the configgin service account token in a secret named configgin
updated by a configgin-helper job. This variable is not defined for pods
already running with a service account bound to the configgin role because
it would make no difference, and more importantly would prevent bootstrapping
of configgin-helper.

[jsc#CAP-828]